fichiers

+++ commands/sondage.js
@@ -3,16 +3,13 @@ var prefix = "?";
 
 
 module.exports.run = async(bot, message, args) => {
-
+    if (!message.member.hasPermission("MANAGE_ROLES")) return message.channel.send("Vous n'avez pas les permissions nécessaires");
     const sondageSlice = message.content.slice(prefix.length + "sondage".length).trim();
 
-    if(message.guild.channels.find('name', '『📊』sondages')){
-        message.delete()
-    }
-
+    message.delete()
     message.delete().catch();
     var embed = new Discord.RichEmbed()
-    .setTitle("**__Sondage du jour !__**")
+    .setTitle("**__Nouveau sondage !__**")
     .setColor("RANDOM")
     .setDescription(sondageSlice)
     .setFooter(`Sondage par ${message.author.tag}`)
